File: Avoid leaving behind file system changes when moveFileTo() fails

+++ modules/juce_core/native/juce_SharedCode_posix.h
@@ -405,6 +405,9 @@ bool File::moveInternal (const File& dest) const
     if (rename (fullPath.toUTF8(), dest.getFullPathName().toUTF8()) == 0)
         return true;
 
+    if (isNonEmptyDirectory())
+        return false;
+
     if (hasWriteAccess() && copyInternal (dest))
     {
         if (deleteFile())
